faculty와 staff 폴더 통합
