+++ app/people/emeritus-faculty/[id]/page.tsx
@@ -5,7 +5,7 @@ import { useParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import useSWR from 'swr';
 
-import { getMockEmeritusFaculty } from '@/apis/faculty';
+import { getMockEmeritusFaculty } from '@/apis/people';
 
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import PeopleImageWithAnimation from '@/components/people/PeopleImageWithAnimation';
@@ -52,7 +52,7 @@ export default function EmeritusFacultyMemberPage() {
         titleMargin="mb-9"
       >
         <div className="flow-root relative mb-10">
-          <PeopleImageWithAnimation showAnimation={showAnimation} imageURL={data.imageURL} />
+          <PeopleImageWithAnimation imageURL={data.imageURL} />
           <div className="break-all">
             <article className="text-neutral-700 font-noto flex flex-col mb-7">
               <h3 className="text-base font-bold leading-8">연락처 정보</h3>
